+++ rick-n-morty-ui/src/components/Main Page/Form/Form.css
@@ -3,9 +3,9 @@
   justify-content: center;
   font-family: "Montserrat";
   padding-top: 4vh;
-  color: black;
-}
 
+}
+/* 
 .form-details {
   display: flex;
 }
@@ -76,4 +76,4 @@
   .form-content {
     width: auto;
   }
-}
+} */
